code refactor for refresh btn
